Bug Fixes for displaying different types of consoles
This update makes changes to the (CS2TP-Group-27-Project-1\ugame-laravel-backend\app\Http\Controllers\ProductsController.php) to make it able to display different types of console types. As long as they have the console key word it can show any type of console.

You will now be able to put any console in and the program will understand where to put it.